Corrections dividende et perf Mementis
Correction de quelques erreurs dans les calculs de perf et de dividendes, ils sont coherents maintenant. Cependant, Il semble y avoir des incoherences dans la brochure entre les formules mathematiques (page 29 par exemple) et les exemple presentés (page 13 par exemple). Cela amene des soucis au niveau du calcul de la perf plafonnée et du payoff

+++ Win32Pricing/src/FCPMementis.hpp
@@ -8,7 +8,7 @@ public:
 	double VLO_;
 
 	/* Point d'entrée */
-	double PE_;
+	float PE_;
 
 	/* Vecteur de performances */
 	PnlVect *performances_;
